+++ extensions/amp-auto-ads/0.1/ping-network-config.js
@@ -1,0 +1,82 @@
+/**
+ * Copyright 2016 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {Services} from '../../../src/services';
+import {buildUrl} from '../../../ads/google/a4a/shared/url-builder';
+import {dict} from '../../../src/utils/object';
+
+/**
+ * A fake ad network integration that is mainly used for testing
+ * and demo purposes. This implementation gets stripped out in compiled
+ * production code.
+ * @implements {./ad-network-config.AdNetworkConfigDef}
+ * @visibleForTesting
+ */
+export class PingNetworkConfig {
+  /**
+   * @param {!Element} autoAmpAdsElement
+   */
+  constructor(autoAmpAdsElement) {
+    this.autoAmpAdsElement_ = autoAmpAdsElement;
+  }
+
+  /** @override */
+  isEnabled() {
+    return true;
+  }
+
+  /** @override */
+  isResponsiveEnabled() {
+    return true;
+  }
+
+  /** @override */
+  getConfigUrl() {
+    return buildUrl(
+      '//lh3.googleusercontent.com/' +
+        'pSECrJ82R7-AqeBCOEPGPM9iG9OEIQ_QXcbubWIOdkY=w400-h300-no-n',
+      {},
+      4096
+    );
+  }
+
+  /** @override */
+  getAttributes() {
+    return dict({
+      'type': '_ping_',
+    });
+  }
+
+  /** @override */
+  getDefaultAdConstraints() {
+    const viewportHeight = Services.viewportForDoc(
+      this.autoAmpAdsElement_
+    ).getSize().height;
+    return {
+      initialMinSpacing: viewportHeight,
+      subsequentMinSpacing: [
+        {adCount: 3, spacing: viewportHeight * 2},
+        {adCount: 6, spacing: viewportHeight * 3},
+      ],
+      maxAdCount: 8,
+    };
+  }
+
+  /** @override */
+  getSizing() {
+    return {};
+  }
+}
